+++ superconfig/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+# [0.11.0-alpha.0](https://github.com/riteshshukla04/react-native-superconfig/compare/v0.10.0...v0.11.0-alpha.0) (2026-02-20)
+
+
+### Bug Fixes
+
+* **ios:** fallback to PATH node when NODE_BINARY is invalid ([#10](https://github.com/riteshshukla04/react-native-superconfig/issues/10)) ([fbd3bcc](https://github.com/riteshshukla04/react-native-superconfig/commit/fbd3bcc584b2b4aee0d39c41c604da8f51f0f82b))
+
+
+### Features
+
+* add-pipelines ([fc8870d](https://github.com/riteshshukla04/react-native-superconfig/commit/fc8870d690b55b793da2b039d48ca0198671b8c1))
+* add-pipelines ([#8](https://github.com/riteshshukla04/react-native-superconfig/issues/8)) ([416f09a](https://github.com/riteshshukla04/react-native-superconfig/commit/416f09ad0ec784fb3deb819440748d616dae1a52))
+
+## [0.10.1](https://github.com/riteshshukla04/react-native-superconfig/compare/v0.10.0...v0.10.1) (2026-02-20)
+
+
+### Bug Fixes
+
+* **ios:** fallback to PATH node when NODE_BINARY is invalid ([#10](https://github.com/riteshshukla04/react-native-superconfig/issues/10)) ([fbd3bcc](https://github.com/riteshshukla04/react-native-superconfig/commit/fbd3bcc584b2b4aee0d39c41c604da8f51f0f82b))
+
+
+### Features
+
+* add-pipelines ([fc8870d](https://github.com/riteshshukla04/react-native-superconfig/commit/fc8870d690b55b793da2b039d48ca0198671b8c1))
+* add-pipelines ([#8](https://github.com/riteshshukla04/react-native-superconfig/issues/8)) ([416f09a](https://github.com/riteshshukla04/react-native-superconfig/commit/416f09ad0ec784fb3deb819440748d616dae1a52))
+
 # [0.10.0](https://github.com/riteshshukla04/react-native-superconfig/compare/v0.9.0...v0.10.0) (2026-02-09)